Revert "Larger tripple num, faster query for distinct predicate"

+++ KVstore/KVstore.cpp
@@ -209,7 +209,7 @@ KVstore::getEntityInDegree(TYPE_ENTITY_LITERAL_ID _entity_id) const
 {
 	//cout << "In getEntityInDegree " << _entity_id << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _entity_id, (char*&)_tmp, _len);
 
 	int ret = 0;
@@ -231,7 +231,7 @@ KVstore::getEntityOutDegree(TYPE_ENTITY_LITERAL_ID _entity_id) const
 {
 	//cout << "In getEntityOutDegree " << _entity_id << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _entity_id, (char*&)_tmp, _len);
 
 	int ret = 0;
@@ -253,7 +253,7 @@ KVstore::getLiteralDegree(TYPE_ENTITY_LITERAL_ID _literal_id) const
 {
 	//cout << "In getLiteralDegree " << _literal_id << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _literal_id, (char*&)_tmp, _len);
 
 	int ret = 0;
@@ -273,7 +273,7 @@ KVstore::getPredicateDegree(TYPE_PREDICATE_ID _predicate_id) const
 {
 	//cout << "In getPredicate Degree " << _predicate_id << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->preID2values, _predicate_id, (char*&)_tmp, _len);
 
 	int ret = 0;
@@ -295,7 +295,7 @@ KVstore::getSubjectPredicateDegree(TYPE_ENTITY_LITERAL_ID _subid, TYPE_PREDICATE
 {
 	//cout << "In getSubjectPredicateDegree " << _subid << ' ' << _preid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _subid, (char*&)_tmp, _len);
 
 	int ret = 0;
@@ -331,7 +331,7 @@ KVstore::getObjectPredicateDegree(TYPE_ENTITY_LITERAL_ID _objid, TYPE_PREDICATE_
 {
 	//cout << "In getObjectPredicateDegree " << _objid << _preid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _objid, (char*&)_tmp, _len);
 
 	int ret = 0;
@@ -534,7 +534,7 @@ bool
 KVstore::updateInsert_s2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID _pre_id, TYPE_ENTITY_LITERAL_ID _obj_id) 
 {
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _sub_id, (char*&)_tmp, _len);
 	bool _is_entity = Util::is_entity_ele(_obj_id);
 
@@ -557,7 +557,7 @@ KVstore::updateInsert_s2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID
 	else 
 	{
 		unsigned* _values;
-		unsigned long _values_len;
+		unsigned _values_len;
 		unsigned _position = KVstore::binarySearch(_pre_id, _tmp + 3, _tmp[1], 2);
 
 		//preID doesn't exist
@@ -641,7 +641,7 @@ bool
 KVstore::updateRemove_s2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID _pre_id, TYPE_ENTITY_LITERAL_ID _obj_id) 
 {
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _sub_id, (char*&)_tmp, _len);
 	bool _is_entity = Util::is_entity_ele(_obj_id);
 
@@ -659,7 +659,7 @@ KVstore::updateRemove_s2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID
 	else 
 	{
 		unsigned* _values;
-		unsigned long _values_len;
+		unsigned _values_len;
 		unsigned _position = KVstore::binarySearch(_pre_id, _tmp + 3, _tmp[1], 2);
 		unsigned _oidlen_sp;
 		//DEBUG: if _tmp[1] -1 < 0??
@@ -763,7 +763,7 @@ bool
 KVstore::updateInsert_o2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID _pre_id, TYPE_ENTITY_LITERAL_ID _obj_id) 
 {
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _obj_id, (char*&)_tmp, _len);
 
 	//objID doesn't exist
@@ -781,7 +781,7 @@ KVstore::updateInsert_o2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID
 	//objID exists
 	else {
 		unsigned* _values;
-		unsigned long _values_len;
+		unsigned _values_len;
 		unsigned _position = KVstore::binarySearch(_pre_id, _tmp + 2, _tmp[1], 2);
 
 		//preID doesn't exist
@@ -861,7 +861,7 @@ bool
 KVstore::updateRemove_o2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID _pre_id, TYPE_ENTITY_LITERAL_ID _obj_id) 
 {
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _obj_id, (char*&)_tmp, _len);
 
 	if (!_get) {
@@ -876,7 +876,7 @@ KVstore::updateRemove_o2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID
 	//objID still exists after removal
 	else {
 		unsigned* _values;
-		unsigned long _values_len;
+		unsigned _values_len;
 		unsigned _position = KVstore::binarySearch(_pre_id, _tmp + 2, _tmp[1], 2);
 		unsigned _sidlen_op;
 		if (_position == _tmp[1] - 1) {
@@ -970,7 +970,7 @@ bool
 KVstore::updateInsert_p2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID _pre_id, TYPE_ENTITY_LITERAL_ID _obj_id) 
 {
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->preID2values, _pre_id, (char*&)_tmp, _len);
 
 	//preid doesn't exist
@@ -985,7 +985,7 @@ KVstore::updateInsert_p2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID
 
 	//preid exists
 	else {
-		unsigned long  _values_len = _len / sizeof(unsigned) + 2;
+		unsigned _values_len = _len / sizeof(unsigned) + 2;
 		unsigned* _values = new unsigned[_values_len];
 		unsigned i, j;
 		_values[0] = _tmp[0] + 1;
@@ -1013,7 +1013,7 @@ bool
 KVstore::updateRemove_p2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID _pre_id, TYPE_ENTITY_LITERAL_ID _obj_id) 
 {
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->preID2values, _pre_id, (char*&)_tmp, _len);
 
 	if (!_get) {
@@ -1027,7 +1027,7 @@ KVstore::updateRemove_p2values(TYPE_ENTITY_LITERAL_ID _sub_id, TYPE_PREDICATE_ID
 
 	//preid still exists after removal
 	else {
-		unsigned long _values_len = _len / sizeof(unsigned) - 2;
+		unsigned _values_len = _len / sizeof(unsigned) - 2;
 		unsigned* _values = new unsigned[_values_len];
 		unsigned i, j;
 		_values[0] = _tmp[0] - 1;
@@ -1694,7 +1694,7 @@ KVstore::getpreIDlistBysubID(TYPE_ENTITY_LITERAL_ID _subid, unsigned*& _preidlis
 		return false;
 	}
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _subid, (char*&)_tmp, _len);
 
 	if (!_get) 
@@ -1731,7 +1731,7 @@ KVstore::getobjIDlistBysubID(TYPE_ENTITY_LITERAL_ID _subid, unsigned*& _objidlis
 		return false;
 	}
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _subid, (char*&)_tmp, _len);
 	if (!_get) 
 	{
@@ -1770,7 +1770,7 @@ KVstore::getobjIDlistBysubIDpreID(TYPE_ENTITY_LITERAL_ID _subid, TYPE_PREDICATE_
 	}
 	unsigned* _tmp = NULL;
 
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _subid, (char*&)_tmp, _len);
 	if (!_get) {
 		_objidlist = NULL;
@@ -1821,7 +1821,7 @@ KVstore::getpreIDobjIDlistBysubID(TYPE_ENTITY_LITERAL_ID _subid, unsigned*& _pre
 	}
 	unsigned* _tmp = NULL;
 
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->subID2values, _subid, (char*&)_tmp, _len);
 	if (!_get) {
 		_preid_objidlist = NULL;
@@ -1984,7 +1984,7 @@ KVstore::getpreIDlistByobjID(TYPE_ENTITY_LITERAL_ID _objid, unsigned*& _preidlis
 {
 	//cout << "In getpreIDlistByobjID " << _objid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _objid, (char*&)_tmp, _len);
 	if (!_get) {
 		_preidlist = NULL;
@@ -2013,7 +2013,7 @@ KVstore::getsubIDlistByobjID(TYPE_ENTITY_LITERAL_ID _objid, unsigned*& _subidlis
 {
 	//cout << "In getsubIDlistByobjID " << _objid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _objid, (char*&)_tmp, _len);
 	if (!_get) {
 		_subidlist = NULL;
@@ -2045,7 +2045,7 @@ KVstore::getsubIDlistByobjIDpreID(TYPE_ENTITY_LITERAL_ID _objid, TYPE_PREDICATE_
 {
 	//cout << "In getsubIDlistByobjIDpreID " << _objid << ' ' << _preid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _objid, (char*&)_tmp, _len);
 	if (!_get) {
 		_subidlist = NULL;
@@ -2085,7 +2085,7 @@ KVstore::getpreIDsubIDlistByobjID(TYPE_ENTITY_LITERAL_ID _objid, unsigned*& _pre
 {
 	//cout << "In getpreIDsubIDlistByobjID " << _objid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->objID2values, _objid, (char*&)_tmp, _len);
 	if (!_get) {
 		_preid_subidlist = NULL;
@@ -2215,7 +2215,7 @@ KVstore::getsubIDlistBypreID(TYPE_PREDICATE_ID _preid, unsigned*& _subidlist, un
 {
 	//cout << "In getsubIDlistBypreID " << _preid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->preID2values, _preid, (char*&)_tmp, _len);
 	if (!_get) {
 		_subidlist = NULL;
@@ -2245,7 +2245,7 @@ bool
 KVstore::getobjIDlistBypreID(TYPE_PREDICATE_ID _preid, unsigned*& _objidlist, unsigned& _list_len, bool _no_duplicate) const {
 	//cout << "In getobjIDlistBypreID " << _preid << endl;
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->preID2values, _preid, (char*&)_tmp, _len);
 	if (!_get) {
 		_objidlist = NULL;
@@ -2279,7 +2279,7 @@ KVstore::getsubIDobjIDlistBypreID(TYPE_PREDICATE_ID _preid, unsigned*& _subid_ob
 	cout << "In getsubIDobjIDlistBypreID " << _preid << endl;
 #endif
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	bool _get = this->getValueByKey(this->preID2values, _preid, (char*&)_tmp, _len);
 #ifdef DEBUG_KVSTORE
 	cout<<"the length of list: "<<_len<<endl;
@@ -2342,7 +2342,7 @@ KVstore::getpreIDlistBysubIDobjID(TYPE_ENTITY_LITERAL_ID _subid, TYPE_ENTITY_LIT
 		return false;
 	}
 	unsigned* _tmp = NULL;
-	unsigned long _len = 0;
+	unsigned _len = 0;
 	this->getValueByKey(this->subID2values, _subid, (char*&)_tmp, _len);
 	_list_len = len;
 	unsigned _result = 0;
@@ -2593,7 +2593,7 @@ KVstore::addValueByKey(IVTree* _p_btree, unsigned _key, char* _val, unsigned _vl
 }*/
 
 bool
-KVstore::addValueByKey(IVArray *_array, unsigned _key, char* _val, unsigned long _vlen)
+KVstore::addValueByKey(IVArray *_array, unsigned _key, char* _val, unsigned _vlen)
 {
 	if (Util::is_literal_ele(_key) && _array == objID2values)
 	{
@@ -2633,7 +2633,7 @@ KVstore::setValueByKey(IVTree* _p_btree, unsigned _key, char* _val, unsigned _vl
 }*/
 
 bool
-KVstore::setValueByKey(IVArray* _array, unsigned _key, char* _val, unsigned long _vlen)
+KVstore::setValueByKey(IVArray* _array, unsigned _key, char* _val, unsigned _vlen)
 {
 	if (Util::is_literal_ele(_key) && _array == objID2values)
 	{
@@ -2674,7 +2674,7 @@ KVstore::getValueByKey(IVTree* _p_btree, unsigned _key, char*& _val, unsigned& _
 }*/
 
 bool
-KVstore::getValueByKey(IVArray* _array, unsigned _key, char*& _val, unsigned long & _vlen) const
+KVstore::getValueByKey(IVArray* _array, unsigned _key, char*& _val, unsigned& _vlen) const
 {
 	if (Util::is_literal_ele(_key) && _array == objID2values)
 	{
@@ -2815,31 +2815,31 @@ KVstore::AddIntoObjCache(TYPE_ENTITY_LITERAL_ID _entity_literal_id)
 	this->objID2values->PinCache(_entity_literal_id);
 }
 
-unsigned long
+unsigned
 KVstore::getSubListSize(TYPE_ENTITY_LITERAL_ID _sub_id)
 {
 	unsigned* _tmp = NULL;
-	unsigned long _ret;
+	unsigned _ret;
 	this->getValueByKey(this->subID2values, _sub_id, (char*&) _tmp, _ret);
 	delete [] _tmp;
 	return _ret;
 }
 
-unsigned long
+unsigned
 KVstore::getObjListSize(TYPE_ENTITY_LITERAL_ID _obj_id)
 {
 	unsigned* _tmp = NULL;
-	unsigned long _ret;
+	unsigned _ret;
 	this->getValueByKey(this->objID2values, _obj_id, (char*&) _tmp, _ret);
 	delete [] _tmp;
 	return _ret;
 }
 
-unsigned long
+unsigned
 KVstore::getPreListSize(TYPE_PREDICATE_ID _pre_id)
 {
 	unsigned* _tmp = NULL;
-	unsigned long _ret;
+	unsigned _ret;
 	this->getValueByKey(this->preID2values, _pre_id, (char*&) _tmp, _ret);
 	delete [] _tmp;
 	return _ret;
